pyproject.toml
[project]
name = "mcp-client-folder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncio>=4.0.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=3.0.3",
    "langchain-mcp-adapters>=0.1.12",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
